+++ sdk/src/main/java/com/hcaptcha/sdk/HCaptchaConfig.java
@@ -121,6 +121,12 @@ public class HCaptchaConfig implements Serializable {
     @Builder.Default
     private long tokenExpiration = 120;
 
+    /**
+     * Enable diagnostics logs that helpful during troubleshooting
+     */
+    @Builder.Default
+    private Boolean diagnosticLog = false;
+
     @JsonIgnore
     @Builder.Default
     private IHCaptchaHtmlProvider htmlProvider = new HCaptchaHtml();