+++ sdk/src/main/java/com/hcaptcha/sdk/HCaptchaHeadlessWebView.java
@@ -27,9 +27,10 @@ final class HCaptchaHeadlessWebView implements IHCaptchaVerifier {
     HCaptchaHeadlessWebView(@NonNull final FragmentActivity activity,
                             @NonNull final HCaptchaConfig config,
                             @NonNull final HCaptchaStateListener listener) {
+        HCaptchaLog.d("HeadlessWebView.init");
         this.config = config;
         this.listener = listener;
-        final WebView webView = new WebView(activity);
+        final WebView webView = new HCaptchaWebView(activity);
         webView.setId(R.id.webView);
         webView.setVisibility(View.GONE);
         if (webView.getParent() == null) {