+++ sdk/src/main/java/com/hcaptcha/sdk/HCaptchaJSInterface.java
@@ -49,6 +49,7 @@ class HCaptchaJSInterface implements Serializable {
 
     @JavascriptInterface
     public void onPass(final String token) {
+        HCaptchaLog.d("JSInterface.onPass");
         handler.post(new Runnable() {
             @Override
             public void run() {
@@ -59,6 +60,7 @@ class HCaptchaJSInterface implements Serializable {
 
     @JavascriptInterface
     public void onError(final int errCode) {
+        HCaptchaLog.d("JSInterface.onError %d", errCode);
         final HCaptchaError error = HCaptchaError.fromId(errCode);
         handler.post(new Runnable() {
             @Override
@@ -70,6 +72,7 @@ class HCaptchaJSInterface implements Serializable {
 
     @JavascriptInterface
     public void onLoaded() {
+        HCaptchaLog.d("JSInterface.onLoaded");
         handler.post(new Runnable() {
             @Override
             public void run() {
@@ -80,6 +83,7 @@ class HCaptchaJSInterface implements Serializable {
 
     @JavascriptInterface
     public void onOpen() {
+        HCaptchaLog.d("JSInterface.onOpen");
         handler.post(new Runnable() {
             @Override
             public void run() {